logcheck: add flag to allow unstructured logs
By default, logcheck will check for use of unstructured logging
and use of incorrect structured logging pattern, and
report errors.

allow-unstructured flag can be passed to logcheck to suppress
errors when unstructured logging is used. It is usefull in
cases when we want to test for correct use of structured
logging patterns but not enfore it.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>