logcheck: support settings in golangci-lint
This implements the new interface from
https://github.com/golangci/golangci-lint/pull/3887 with two fields in the
"settings":

type settings struct {
      Check  map[string]bool `json:"check"`
      Config string          `json:"config"`
}

This is the actual config *data*, not the name of a config file. This is
intentional because
- relative file paths wouldn't work (plugin cannot resolve them)
- loading additional file content wouldn't invalidate cached linter
  results the that embedding the data inside the main config does

The `include` text template method ensures that logcheck.conf gets injected
automatically. Having it as separate file has the advantage that manual
invocations of the stand-alone logcheck linter still work.