logcheck: support banning non-contextual functions and methods
Sometimes there are APIs with different variants where one variant is fine in
code which doesn't support contextual logging but should be replaced with some
other variant if that is the goal.

For example, k8s.io/apimachinery/pkg/util/runtime.HandleError is the old-style
function. We cannot deprecate it because it works fine in most existing
code. But the upcoming HandleErrorWithContext is what code in Kubernetes that
has (or will be) converted to contextual logging should use.

We also cannot forbid it through forbidigo, because then the logcheck.conf with
its definition of which code is "contextual" would be ignored. Besides,
maintaining a config for forbidigo would be continual effort that would need to
be replicated by other projects consuming the package with these APIs.

A better solution is to:
- mark these non-contextual APIs through a `//logcheck:context // <why>` comment
- detect those comments in logcheck
- warn about using them in contextual code