go.mod: bump to Go 1.22
The newer version is required when checking packages which need Go 1.22.
From a csi-lib-utils CI run with Go 1.22.2 in the PATH and Kubernetes 1.30
in vendor:

   /home/prow/go/src/github.com/kubernetes-csi/csi-lib-utils/vendor/k8s.io/apimachinery/pkg/api/resource/amount.go:17:1: package requires newer Go version go1.22

This feels like a workaround and raises the question whether this has to be
done for each new Go release. Discussed in
https://github.com/golang/go/issues/67282.

It is unclear whether building with older Go releases still works. The GitHub
workflow failed while installing the older Go, without any error message.